backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
asyncpg>=0.30.0
psycopg2-binary>=2.9.9

# HTTP client
httpx>=0.28.0

# CLI
click>=8.1.0

# Data processing
pydantic>=2.10.0
python-dotenv>=1.0.0
